+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Change Log
 <a name="v0.0.79"></a>
 ## Version [v0.0.79](https://github.com/luizfp/PrologWebService/compare/v0.0.78...v0.0.79) (release-date) [unreleased]
 
+### Features
+* Cria relatório do cronograma das aferições de placas
+
 ### Refactors
 * Adiciona infos extras no gráfico de scatter dos pneus com menor sulco e pressão da dashboard